requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
notion-client==2.2.1
sentence-transformers==2.7.0
numpy==1.26.4
huggingface-hub==0.23.2
llama-cpp-python==0.2.76
